There should be no unnecessary indentations in HTML
When I first started to work on this website I think I used
[Brackets][1] editor which had 4 spaces as its default indentation
setting for HTML and CSS.

Later, when I moved editing of this project to Emacs, I was frustrated
by Emacs putting two spaces instead.

So I decided to google if there's a common convention for it to
[get this finally straight][2] once and for all.

Turns out, w3schools.com (which is a sort of authority for me when it
comes to web development) discourages any indentation that doesn't have
a reason for it!

[1]: http://brackets.io/
[2]:
https://stackoverflow.com/questions/5430806/html-indenting-standard-tab-or-two-spaces

PS I guess, I'll have to reformat my website completely now at some
point later on :sweat_smile: